move factorise_integer function to algorithm module
