+++ pyrate/configuration.py
@@ -18,7 +18,7 @@ from pathlib import Path, PurePath
 import re
 from pyrate.constants import NO_OF_PARALLEL_PROCESSES
 from pyrate.default_parameters import PYRATE_DEFAULT_CONFIGURATION
-from pyrate.core.user_experience import factorise_integer
+from pyrate.core.algorithm import factorise_integer
 from pyrate.core.shared import extract_epochs_from_filename
 from pyrate.core.config import parse_namelist, ConfigException
