+++ pyrate/core/algorithm.py
@@ -235,3 +235,42 @@ def master_slave_ids(dates):
 
     dset = sorted(set(dates))
     return dict([(date_, i) for i, date_ in enumerate(dset)])
+
+
+def factorise_integer(n, memo={}, left=2):
+    """
+    Returns two factors a and b of a supplied number n such that a * b = n.
+    The two factors are evaluated to be as close to each other in size as possible
+
+    :param int n: Number to factorise
+    :param dict memo: dictionary of candidate factors
+    :param int left: operation flag (default = 2)
+
+    :return: a, factor one
+    :rtype: int
+    :return: b, factor two
+    :rtype: int
+    """
+    if (n, left) in memo:
+        return memo[(n, left)]
+    if left == 1:
+        return n, [n]
+    i = 2
+    best = n
+    bestTuple = [n]
+    while i * i <= n:
+        if n % i == 0:
+            rem = factorise_integer(n / i, memo, left - 1)
+            if rem[0] + i < best:
+                best = rem[0] + i
+                bestTuple = [i] + rem[1]
+        i += 1
+
+    # handle edge case when only one processor is available
+    if bestTuple == [4]:
+        return 2, 2
+
+    if len(bestTuple) == 1:
+        bestTuple.append(1)
+
+    return int(bestTuple[0]), int(bestTuple[1])
